st_app/requirements.txt
Pillow==11.2.1
Requests==2.32.4
streamlit==1.45.1
